requirements.txt
streamlit==1.28.0
openai==1.3.0
sympy==1.12
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.24.3
Pillow==10.1.0
streamlit-drawable-canvas==0.9.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pytesseract==0.3.10
opencv-python==4.8.1.78
